fixed all the scripts

+++ scripts/cassandra/benchmark.sh
@@ -1,9 +1,10 @@
 #initialize the database
 duration=1m
+pids=$(pgrep -f cassandra | paste -s -d,)
 sudo cassandra-stress write n=1000000 -rate threads=50 > ./write_output.txt
 
 #first test : 3,7
-pids=$(pgrep -f cassandra | paste -s -d,)
+
 echo -e $pids
 sudo cassandra-stress mixed ratio\(write=3,read=7\) duration=$duration -rate threads=50 >./output-3-7.txt 2>&1 &
 sudo strace -f -p "$pids" 2> ./benchmark/cassandra-benchmark-3-7-bare-metal-$(uname -r).strace &
@@ -12,10 +13,17 @@ sudo perf record -p $pids -o ./benchmark/cassandra-benchmark-3-7-bare-metal-$(un
 perf_id=$!
 echo $perf_id
 sleep $duration
-kill $perf_id
-kill $strace_id
+# Check if strace process is still running before trying to kill it
+if ps -p $strace_id > /dev/null; then
+   kill $strace_id
+fi
+
+# Check if perf process is still running before trying to kill it
+if ps -p $perf_id > /dev/null; then
+   kill $perf_id
+fi
 echo "done first test"
-sleep 10
+sleep 20
 
 #second test : 5,5
 echo -e $pids
@@ -26,10 +34,17 @@ sudo perf record -p $pids -o ./benchmark/cassandra-benchmark-5-5-bare-metal-$(un
 perf_id=$!
 echo $perf_id
 sleep $duration
-kill $perf_id
-kill $strace_id
+# Check if strace process is still running before trying to kill it
+if ps -p $strace_id > /dev/null; then
+   kill $strace_id
+fi
+
+# Check if perf process is still running before trying to kill it
+if ps -p $perf_id > /dev/null; then
+   kill $perf_id
+fi
 echo "done second test"
-sleep 10
+sleep 20
 
 #third test 7,3
 sudo cassandra-stress mixed ratio\(write=7,read=3\) duration=$duration -rate threads=50 >./output-7-3.txt 2>&1 &
@@ -39,10 +54,18 @@ sudo perf record -p $pids -o ./benchmark/cassandra-benchmark-7-3-bare-metal-$(un
 perf_id=$!
 echo $perf_id
 sleep $duration
-kill $perf_id
-kill $strace_id
+# Check if strace process is still running before trying to kill it
+if ps -p $strace_id > /dev/null; then
+   kill $strace_id
+fi
+
+# Check if perf process is still running before trying to kill it
+if ps -p $perf_id > /dev/null; then
+   kill $perf_id
+fi
 echo "done Third test"
-sleep 10
+sleep 20
+
 # # sudo perf record -o ./benchmark/cassandra-benchmark-3-7-bare-metal-$(uname -r).perf -e tlb:tlb_flush,dTLB-loads,dTLB-load-misses,iTLB-load-misses,cache-misses,page-faults cassandra-stress write n=1000000 -rate threads=50
 
 # sudo perf record -o ./benchmark/cassandra-benchmark-3-7-bare-metal-$(uname -r).perf -e tlb:tlb_flush,dTLB-loads,dTLB-load-misses,iTLB-load-misses,cache-misses,page-faults cassandra-stress mixed ratio\(write=3,read=7\) n=1000000 -rate threads=50
